rmvar: mv EntityMap implementation -> IO.inl
This isn't exactly how to make C++ look simple :(

It's a separate class, though just a helper to IO. I guess according to
the coding style, it have it's enirely separate set of source files,
but that doesn't seem right. I guess I could make it a class within
IO, would that be preferred?